BUG fix:  layer and sector indices were swapped in generation of dark hits causing fewer than normal dark hits.
Also I have removed scaling the thresholds by the square root of the number of cells and reverted by to scaling by the number of cells.  Scaling by the square root, which seems natural for Gaussian statistics, produces a huge number of spurious showers due to dark hits above threshold.  This clearly needs some more careful study to understand how to properly set the threshold of the sum of channels.