Majorly revised BCAL smearing code to use timing spectra from hdgeant rather than integrated hit info.
NOTE: The changes made were placed inside a preprocessor directive that 
currently is set to compile the old smearing scheme and not the new one
described above. One should therefore see no functional change between this revision and the last unless the "#if 0" at the top of smear_bcal is changed to "#if 1".

The new code runs about 15 times slower and uses histogram binning that is 1ns wide. This will not be sufficient to achieve the ~50-100ps resolution expected (see https://halldweb1.jlab.org/wiki/images/6/63/20111031_bcal.pdf). Decreasing bin width will slow it down much further so additional profiling and optimization will be needed.