Reverting BCAL simulation back to original version as the default. This is in response to Paul's request since the reconstruction is not yet ready to handle the new scheme.
The change is not a complete reversal though:

1. Both types of BCAL data are now written out by default from the
BCAL in hdgeant.

2. Which method is used (new or old) in mcsmear is chosen by
command line parameter (-D) with the default being the old.

3. The DBCALPoint object no longer scales the signal amplitude
to convert from ADC counts to rough GeV since that is not
needed by the old scheme. This must be changed in the code
and recompiled to get the "new" behavior.

4. Both mcsmear and all DANA programs will print an annoying 
warning message to remind users that they are using a deprecated
method.