After some discussion with Mark, backed out changes from the last several weeks aimed at supressing warnings. We agreed that it would be too much effort (and too dangerous) to modify pythia to the point that it would compile cleanly with -Wall. This was done using a reverse merge using the following command:
svn merge -c -XXX https://halldsvn.jlab.org/repos/trunk/sim-recon/src/programs/Simulation/bggen/code

where "XXX" was each of the following revisions in this order:

12238 (pythia_h.F only)
12029
11933
11880
11877
11876
11849
11782
11728

Other changes committed in between these revisions should still be
there. This includes the changes that remove -Wall when compiling
the Fortran code in bggen so that it does not emit a long list of
warnings.