From wlevine:
Temporary fix to mcsmear to suppress large number of BCAL dark hits
not seen in real data.

This disables dark hit generation in BCAL cells which do not have any
true signal in them. The reason we can't disable dark hit generation
entirely is that this would have significant unintended effects on
simulation of real hits.